improving flush, adding onAllDataFlushed
1. flush : when calling twice one after another ( or when flush called
manually directly follows a flush triggered internally ), the second
flush would trigger its callback before the previous could finish
sending its data. This is misleading since you would usually expect
data data you have just enqueued, has been safely uploaded once the
flush callback is called. The commit improves this behaviour by waiting
for currently pending flush to finish before executing the callback of
the consecutive flush ( as long as there was no events in the queue when
the second flush was called ).
2. onAllDataFlushed : even with the above improvement, if we rally want
to be sure that there is no pending data ( for example if we want to
close the app ), we need to do a little bit more work, since even though
a flush is done, it doesn’t mean that new data has not been queued and
flushed in the meantime ( between a call to flush and it callback ).
onAllDataFlushed registers a callback ( listener ) which fires once,
whenever a flush is done and there are no more messages in the queue.