Stringify userId and anonymousId
Ref: https://segment.atlassian.net/browse/LIB-29

Our API truncates bignum values. This is problematic because this library has historically accepted Numbers as userIds and anonymousIds. To workaround the truncation, this library needs to stringify userId and anonymousId on the client. In the future we can make a breaking change and restrict the library to only accepting String ids, but for now we are keeping backward compatibility.