Remove RUN_E2E_TESTS from Circle config.
Instead of setting this up via the config file, we'll do this as an encrypted variable via the UI https://circleci.com/gh/segmentio/analytics-node/edit#env-vars.

Otherwise the RUN_E2E_TESTS variable is true for contributors, but they won't have the `RUNSCOPE_TOKEN` which is encrypted and not available to them.

This causes contributor PRs to fail as the test suite tries to run the end to end tests without the appropriate tokens.

LIB-176 - https://segment.atlassian.net/browse/LIB-176