Ensure callback is called when envoking errorHanlder
A property, errorHanlder, was recently added which will be called
instead of throwing an error in the .flush() method. This is important
because errors in .flush() could, previously, only be handled via
process.on('uncaughtException', err => { ... }).

However, this property is currently unusable as, when the flush method
invokes this property, it fails to call the callbacks of the events
being flushed.

This commit makes sure the callbacks are called.